+++ etc/sync-function-template.js
@@ -440,13 +440,7 @@ function synctos(doc, oldDoc) {
           validateArray(doc, oldDoc, validator.arrayElementsValidator, itemStack, validationErrors);
           break;
         case 'hashtable':
-          validateHashtable(
-            doc,
-            oldDoc,
-            validator.hashtableKeysValidator,
-            validator.hashtableValuesValidator,
-            itemStack,
-            validationErrors);
+          validateHashtable(doc, oldDoc, validator, itemStack, validationErrors);
           break;
         case 'attachmentReference':
           validateAttachmentRef(doc, oldDoc, validator, itemStack, validationErrors);
@@ -614,20 +608,24 @@ function synctos(doc, oldDoc) {
     }
   }
 
-  function validateHashtable(doc, oldDoc, keyValidator, valueValidator, itemStack, validationErrors) {
+  function validateHashtable(doc, oldDoc, validator, itemStack, validationErrors) {
+    var keyValidator = validator.hashtableKeysValidator;
+    var valueValidator = validator.hashtableValuesValidator;
     var currentItemEntry = itemStack[itemStack.length - 1];
     var itemValue = currentItemEntry.itemValue;
     var oldItemValue = currentItemEntry.oldItemValue;
+    var hashtablePath = buildItemPath(itemStack);
 
     if (typeof itemValue !== 'object' || itemValue instanceof Array) {
       validationErrors.push('item "' + buildItemPath(itemStack) + '" must be an object/hashtable');
     } else {
+      var size = 0;
       for (var elementKey in itemValue) {
+        size++;
         var elementValue = itemValue[elementKey];
 
         var elementName = '[' + elementKey + ']';
         if (keyValidator) {
-          var hashtablePath = buildItemPath(itemStack);
           var fullKeyPath = hashtablePath ? hashtablePath + elementName : elementName;
           if (typeof elementKey !== 'string') {
             validationErrors.push('hashtable key "' + fullKeyPath + '" is not a string');
@@ -662,6 +660,14 @@ function synctos(doc, oldDoc) {
 
           itemStack.pop();
         }
+      }
+
+      if (!isValueNullOrUndefined(validator.maximumSize) && size > validator.maximumSize) {
+        validationErrors.push('hashtable "' + hashtablePath + '" must not be larger than ' + validator.maximumSize + ' elements');
+      }
+
+      if (!isValueNullOrUndefined(validator.minimumSize) && size < validator.minimumSize) {
+        validationErrors.push('hashtable "' + hashtablePath + '" must not be smaller than ' + validator.minimumSize + ' elements');
       }
     }
   }