+++ test/resources/hashtable-doc-definitions.js
@@ -1,0 +1,15 @@
+{
+  hashtableDoc: {
+    channels: { write: 'write' },
+    typeFilter: function(doc) {
+      return doc._id === 'hashtableDoc';
+    },
+    propertyValidators: {
+      sizeValidationProp: {
+        type: 'hashtable',
+        minimumSize: 2,
+        maximumSize: 2
+      }
+    }
+  }
+}